requirements.txt
asgiref==3.8.1
attrs==23.2.0
Babel==2.14.0
certifi==2024.6.2
chardet==5.2.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
Django==5.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
filelock==3.15.4
idna==3.6
inflect==7.3.1
Jinja2==3.1.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.19.2
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
more-itertools==10.3.0
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
PyJWT==2.7.0
PyMySQL==1.1.1
python-dateutil==2.9.0
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.0.7
virtualenv==20.26.2
wheel==0.44.0
